% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_first_id.R
\name{extract_first_id}
\alias{extract_first_id}
\title{Extract the first string of each element containing delimiter-separated concatenated strings (in a character vector)}
\usage{
extract_first_id(concatstr, splitchar = ";")
}
\arguments{
\item{concatstr}{character vector, each element of which is to be split and first element extracted from}

\item{splitchar}{character object to use for splitting by \code{strsplit}, default ";"}
}
\value{
character vector containing first string of each element of \code{concatstr}
}
\description{
Extract the first string of each element containing delimiter-separated concatenated strings (in a character vector)
}
\examples{
library(SY351SILAC)
library(tidyverse)
psitestats$numgenes <- unlist(lapply(strsplit(psitestats$Gene.names, ";"), length))
table(psitestats$numgenes)
concatids <- psitestats \%>\% dplyr::filter(numgenes > 1) \%>\% dplyr::slice(1:20) \%>\% dplyr::select(Gene.names) \%>\% dplyr::pull()
cbind(extract_first_id(concatids, ";"),concatids)
concatids[6] <- "MYC"
concatids[2] <- ""
cbind(extract_first_id(concatids, ";"),concatids)

}
