% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protein_meta_data.R
\name{get_protein_meta_data}
\alias{get_protein_meta_data}
\title{Loads protein-level meta data from a Perseus annotated phosphosite MaxQuant file.}
\usage{
get_protein_meta_data(pmetafile, absfilepath, is_gzip_file = T)
}
\arguments{
\item{pmetafile}{character, name of protein meta data file}

\item{absfilepath}{character, file path}

\item{is_gzip_file}{logical, is the file gzipped?, default TRUE}
}
\value{
data frame (\code{tibble}) with the protein meta data and an additional column \code{proteinclass} annotating whether it is a kinase
}
\description{
This assumes that the file \code{pmetafile} has been previously generated using
Perseus by annotating a maxQuant PhosphoSite file (Phospho(STY).txt usually), with protein-level
and phosphosite-level annotations. Remove the data columns to avoid duplicate columns in downstream
merges with data column containing data frames or tibbles. The file must have the following columns
 'Unique identifier': key to merge with phosphosite data table.
 'Interpro.name': This and next two columns are used to annotate whether the protein is a kinase
 'Prosite.name'
 'Pfam'
}
\examples{

}
