% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_toptable.R
\name{get_toptable}
\alias{get_toptable}
\title{Table of statistics from an eBayes differential analysis object}
\usage{
get_toptable(fitobj, coef, glist, suffix)
}
\arguments{
\item{fitobj}{list containing a linear model fit produced by lmFit and eBayes.
fit should be an object of class MArrayLM as produced by lmFit and eBayes.}

\item{coef}{name of coefficient in \code{fitobj}}

\item{glist}{dataframe of annotation columns from the original data used to create \code{fitobj}}

\item{suffix}{suffix to append to the statistic columns}
}
\value{
A data frame derived from a call to limma's topTable() function, with s2.prior and sigma from the lmFit and eBayes calls.
}
\description{
Extract a table containing the test statistics, p-values, adjusted p-values, etc. from
a linear model fit produced by lmFit and eBayes from the limma package. A suffix can be
provided to append to the column names for the statistics, to allow for downstream workflows.
}
\examples{

}
