% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvolcano.R
\name{ggvolcano}
\alias{ggvolcano}
\title{Volcano plot of differential proteomic data with log2 ratios, p-values and adjusted p-values.}
\usage{
ggvolcano(
  tdf,
  xvar,
  yvar,
  labelvar,
  adjpvalvar,
  fdr.range.cut = c(0, 0.01, 0.05, 0.1, 1),
  high.sig.threshold = 0.1,
  xlimits = Inf,
  ylimits = Inf,
  xtitle = NULL,
  ytitle = NULL,
  plottitle = NULL,
  annotgenes = T,
  revcols = T
)
}
\arguments{
\item{tdf}{A data frame containing the log2 ratios (xvar), p-values (yvar), labeling variable (labelvar)}

\item{xvar}{x-axis bare variable name of numeric column in tdf data frame, usually log2 ratio}

\item{yvar}{y-axis bare variable name of numeric column in tdf data frame, usually log10(unadjusted p-value)}

\item{labelvar}{column in tdf (bare variable name) to use for labeling individual points}

\item{adjpvalvar}{column in tdf (bare variable name) to use for adjusted p-value}

\item{fdr.range.cut}{numeric vector, FDR intervals to indicate significance and for mapping color, sorted from smallest (zero) to largest (1).}

\item{high.sig.threshold}{numeric, all points with adj.p.value <= to this are annotated with text specified by labelvar}

\item{xlimits}{length-2 numeric, x-axis limits}

\item{ylimits}{length-2 numeric, y-axis limits}

\item{xtitle}{character, x-axis title}

\item{ytitle}{character, y-axis title}

\item{plottitle}{character, plot title}

\item{annotgenes}{logical, should genes with adjusted p-value < high.sig.threshold be explicitly labeled}

\item{revcols}{logical, reverse the color vector}
}
\value{
a named list with two plot objects, both of which can be individually plotted by assigning
    to a new variable and typing variable name at command line:
      ggplot_volcano: a ggplot version of volcano plot
      plotly_volcano: an interactive plotly volcano plot.
}
\description{
\code{ggvolcano} returns a list with two volcano plot objects (ggplot2 and plotly).
The data frame (parameter tdf) contains the log2 ratios (xvar), p-values (yvar),
adjusted p-values (adjpvalvar), labeling variable (labelvar). The columns of tdf that
correspond to each of these variables are provided to the ggvolcano function as unquoted
column names (in the tidyverse fasion) for the corresponding variables in the tdf data frame.
}
\examples{
library(SY351SILAC) # loads the psitestats data frame
library(tidyverse)
psitestats <- psitestats \%>\% dplyr::mutate(logp = -log10(P.Value.nullfiltisotope))

mygglist <- ggvolcano(psitestats, xvar =  logFC.nullfiltisotope,
                      yvar = logp,
                      fdr.range.cut = c(0,0.01,1),
                      adjpvalvar = adj.P.Val.nullfiltisotope,
                      labelvar = gene.psite,
                      high.sig.threshold = 0.001,
                      xlimits = c(-3.5,3.5),
                      ylimits = c(0,8),
                      annotgenes = T,
                      xtitle = 'log2(SY351/DMSO)',
                      ytitle = 'log10(p-value)')


mygglist$ggplot_volcano

mygglist$plotly_volcano

}
