% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itsar.R
\docType{package}
\name{itsar}
\alias{itsar}
\alias{itsar-package}
\title{Identification of drug-binding targets from iTSA thermal profiling data}
\description{
R A package with functions for importing proteomics data (MaxQuant output files),
differential statistical analysis, and visualization of differential proteomics data,
particularly for thermal profiling data collected using the iTSA method,
Ball KA, Webb KJ, Coleman SJ, Cozzolino KA, Jacobsen J, Jones KR, Stowell MHB, Old WM (2020).
An isothermal shift assay for proteome scale drug-target identification. Communications Biology 3, 1–10.
DOI: 10.1038/s42003-020-0795-6
}
\details{
itsar
A package for proteome-wide identification of drug targets from iTSA thermal profiling data
}
