% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_maxquant_proteingroupfile.R
\name{read_maxquant_proteingroupfile}
\alias{read_maxquant_proteingroupfile}
\title{Read protein quantification table from MaxQuant output file proteinGroups.txt}
\usage{
read_maxquant_proteingroupfile(
  proteinGroups.filename = "proteinGroups.txt",
  datacolnames = NULL,
  datacolumnprefix = "LFQ intensity ",
  uniprot.acc.column = "Majority protein IDs",
  uniprot.id.column = "UniProt names",
  meta.colnms = c("Protein.names", "Gene.names", "uniprot.ids", "uniprot.accs",
    "Isoform.ID", "Sequence.lengths"),
  null.zeros = TRUE,
  log2.transform = T,
  is.gzip.file = F
)
}
\arguments{
\item{proteinGroups.filename}{file name string, assuming it is located in current directory. Default "proteinGroups.txt"}

\item{datacolnames}{character array of the data column names to extract, which will ignore the datacolumnprefix parameter, optional}

\item{datacolumnprefix}{string used to identify the column name containing the quantification. Default "LFQ intensity "}

\item{uniprot.acc.column}{column with the uniprot accessions (protein identifiers), default "Majority protein IDs"}

\item{uniprot.id.column}{column with the uniprot entry name, (previously called uniprot IDs, default "UniProt names"}

\item{meta.colnms}{character array of additional columns to include}

\item{null.zeros}{logical, If TRUE, convert zero intensities to NA. Default TRUE}

\item{log2.transform}{logical, log2 transform the intensities? Default TRUE}

\item{is.gzip.file}{boolean indicating whether the file is gzipped (*.gz)}
}
\value{
a list with the following elements: prot.table (the data table) and
    prot.colnms (the data column names indexed with datacolumnprefix parameter)
}
\description{
Read protein quantification table from a file derived from the MaxQuant output file proteinGroups.txt,
which may have been annotated using the Perseus program.
one row per protein, with samples represented in columns as specified in the
MaxQuant experimental design file summary.table. Be careful with SILAC  - FBS
contamination will result in unlabled peptides identical to human orthologs and
will inflate the light channel,and this function filters contaminants out. Note
that this function (read.table actually) will use make.names during the import of
the proteinGroups.txt file and will convert any column names in that file by using
make.names, and will for example convert spaces to "." in the resulting prot.table
}
\examples{
read_maxquant_proteingroupfile()
}
