% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silac_model_matrix.R
\name{silac_model_matrix}
\alias{silac_model_matrix}
\title{Create model matrix for a two label isotope-labeled SILAC design, given a targets data frame}
\usage{
silac_model_matrix(targets, parameters = NULL, ref = NULL, verbose = TRUE)
}
\arguments{
\item{targets}{matrix or data frame with columns \code{Light} and \code{Heavy} describing which samples are labeled with which isotope}

\item{parameters}{Not used. matrix specifying contrasts between RNA samples which should correspond to regression coefficients. Row names should correspond to unique RNA sample names found in targets.}

\item{ref}{character string giving name of one of the RNA sources to be treated as reference. Exactly one argument of parameters or ref should be specified.}

\item{verbose}{logical, if TRUE then unique names found in targets will be printed to standard output}
}
\value{
produces a numeric design matrix with row names as in targets and column names as in parameters.
}
\description{
Modeled after limma's \code{\link[limma]{modelMatrix}} function
 Make sure the the target data frame has two columns: Light and Heavy
 which indicates the sample's SILAC label.
 targets. e.g. for a label-swap design:
 samples   Light   Heavy
 1    null  DMSO  DMSO
 2    rep1  DMSO SY351
 3    rep2  DMSO SY351
 4  rep3LF SY351  DMSO
}
\examples{
targets <- data.frame(samples = c('null','rep1','rep2','rep3LF'),
                      Light = c("DMSO","DMSO","DMSO","SY351"),
                      Heavy  = c("DMSO","SY351","SY351","DMSO"),
                      samplenames = c("Ratio.H.L.normalized.null",
                                     "Ratio.H.L.normalized.rep1",
                                     "Ratio.H.L.normalized.rep2",
                                     "Ratio.H.L.normalized.rep3LF"))
design <- silac_model_matrix(targets, ref = "DMSO")
design
design.isotopeffect <- cbind(rep(1,length(targets$samples)), design)
colnames(design.isotopeffect) <- c("isotope","SY351")
design.isotopeffect

}
